Update test decks to be more evergreen
Uses stable cards like vanillas for unlimited and an exodia piece for limited, instead of the volatile status of an ABC deck. The downside is Left Arm has the same status in both regions, as would all stable options barring arguably Maxx "C", making it impossible to have region-specific tests.