Let the allow vector define what aliases are applied (#111)
The vector contains all cards that will be considered. Cards not in it are automatically illegal. To allow an alias and treat it as the main, set a negative value for it. If an alias is set to a positive value then it's treated as a separate card. Caveat: if the main is forbidden and some aliases are to be allowed, they cannot be treated as separate cards.